% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{try_register}
\alias{try_register}
\title{Test APPG availability}
\usage{
try_register()
}
\value{
A boolean indicating whether we can use static parsing
}
\description{
Test APPG availability
}
\examples{
 result <- try_register()
}
